fix: set user data after registration
