+++ src/components/Register/Register.scss
@@ -6,11 +6,3 @@
     align-items: center;
     text-align: center;
 }
-
-.field {
-    margin: 5vh;
-}
-
-.response-message {
-    margin-top: 5vh;
-}
