+++ src/components/Register/RegisterForm/RegisterForm.scss
@@ -1,0 +1,7 @@
+.field {
+    margin: 5vh;
+}
+
+.response-message {
+    margin-top: 5vh;
+}
